pyproject.toml
[project]
name = "image-semantic-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-experimental>=0.3.4",
    "loguru>=0.7.3",
    "matplotlib>=3.10.1",
    "open-clip-torch>=2.32.0",
    "pillow>=11.2.1",
    "scikit-learn>=1.6.1",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
]
